Move the 'Ingredient' class to a separate file
